Adição ResultByClasses e alteração na main.py

+++ src/main.py
@@ -14,8 +14,10 @@ def train_model(model,fold,fold_dir,ROOT_DATA_DIR):
     check_save_path = os.path.join(fold_dir,model)
 
     if os.path.exists(check_save_path):
+        if CONTINUE:
+            return None
         shutil.rmtree(check_save_path)
-        
+
     if model == 'YOLOV8':
         from Detectors.YOLOV8.RunYOLOV8 import runYOLOV8
         runYOLOV8(fold,fold_dir,ROOT_DATA_DIR)
@@ -30,7 +32,6 @@ def train_model(model,fold,fold_dir,ROOT_DATA_DIR):
         from Detectors.Detr.runDetr import runDetr
         runDetr(fold,fold_dir,ROOT_DATA_DIR)
         model_path = os.path.join(fold_dir,model,'training','best_model.pth')
-
     return model_path
 # Função que server para selecionar os modelos que ja foram treinados
 def test_model(model,fold_dir):
@@ -45,16 +46,16 @@ def test_model(model,fold_dir):
     return model_path
 
 # YOLOV8, Faster, Detr
-MODELS = ['YOLOV8','Faster','Detr','sabl'] #Variavel para selecionar os modelos
+MODELS = ['YOLOV8'] #Variavel para selecionar os modelos
 
-APENAS_TESTE = True # True para apenas testar modelos treinados False para Treinar e Testar.
+APENAS_TESTE = False # True para apenas testar modelos treinados False para Treinar e Testar.
 ROOT_DATA_DIR = os.path.join('..', 'dataset','all')
 DIR_PATH = os.path.join(ROOT_DATA_DIR, 'filesJSON')
 DOBRAS = int(len(os.listdir(DIR_PATH))/3)
 GeraRult = True # True para gerar Resultados False para não gerar
 save_imgs = True # True para salvar imagens em predictes False para não salvar
-GeraResultByClass = True
-
+GeraResultByClass = True # True para Salvar Resultados Por classes
+CONTINUE = False # True para Continuar sem apagar os pesos ja treinados
 resetar_pasta(os.path.join("..","results","prediction"))
 
 if GeraRult:
@@ -76,6 +77,8 @@ for model in MODELS:
         fold_dir = os.path.join('model_checkpoints', fold)
         if not APENAS_TESTE:
             model_path = train_model(model,fold,fold_dir,ROOT_DATA_DIR)
+            if model_path == None:
+                continue
         else:
             model_path =  test_model(model,fold_dir)
